Add files via upload
The project, Feel Well Web App, is a web development project. It is a mental well-being portal where we aim to spread awareness and help users find ways to improve their mental health. It is designed with a minimalist interface for a better user experience without any hurdles.

The project consists of 3 main pages: the home page, the blog page, and the HappySpace page. The home page is the main page of the portal, comprising a brief description of the website, contact details, testimonials, and other navigation options to move anywhere inside the portal.

The blog page is designed for the purpose of sharing articles and other pieces of writing about mental health awareness and some tips and tricks via interactive blogs, so that users can learn to prioritise their mental health.

The HappySpace page is designed with the thought of how can improve user's mood and help them have a change of mind and be relaxed. Four branches have been added to it